chore: move TokenService to util folder
